fix interop issues.

+++ src/quic/crypto/if_cryptographer.h
@@ -56,13 +56,22 @@ public:
     virtual Result EncryptPacket(uint64_t pn, common::BufferSpan& associated_data, common::BufferSpan& plaintext,
         std::shared_ptr<common::IBuffer> out_ciphertext) = 0;
 
-    virtual Result DecryptHeader(common::BufferSpan& ciphertext, common::BufferSpan& sample, uint8_t pn_offset,
+    // `pn_offset` is the distance from the start of the packet to the packet
+    // number field. It must be at least 32 bits wide: a long header carries two
+    // connection IDs of up to 20 B each and an Initial sent in response to a
+    // Retry also echoes the server's token, which is commonly 256 B (aioquic),
+    // so the offset routinely exceeds 255. Narrowing it to uint8_t silently
+    // wraps the offset modulo 256 and applies the header-protection mask deep
+    // inside the token instead of the packet number, corrupting the token and
+    // leaving the packet number unprotected.
+    virtual Result DecryptHeader(common::BufferSpan& ciphertext, common::BufferSpan& sample, uint32_t pn_offset,
         uint8_t& out_packet_num_len, bool is_short) = 0;
 
     // RFC 9001 §6: Check if previous read key is available (for Key Update fallback)
     virtual bool HasPrevReadKey() const = 0;
 
-    virtual Result EncryptHeader(common::BufferSpan& plaintext, common::BufferSpan& sample, uint8_t pn_offset,
+    // See DecryptHeader above for why `pn_offset` must not be narrowed.
+    virtual Result EncryptHeader(common::BufferSpan& plaintext, common::BufferSpan& sample, uint32_t pn_offset,
         size_t pkt_number_len, bool is_short) = 0;
 
     virtual size_t GetTagLength() = 0;
